refactor: Move ModelSpec from tests to core library
