+++ cpp/tensorrt_llm/pybind/CMakeLists.txt
@@ -15,6 +15,7 @@ set(SRCS
     executor/executorConfig.cpp
     executor/request.cpp
     runtime/bindings.cpp
+    testing/modelSpecBinding.cpp
     userbuffers/bindings.cpp
     ../runtime/ipcNvlsMemory.cpp
     bindings.cpp)
